+++ Sources/AriaFlow/AriaFlowApp.swift
@@ -117,7 +117,6 @@ struct AriaFlowApp: App {
         Settings {
             SettingsWindowView()
                 .environmentObject(store)
-                .frame(width: 400, height: 360)
                 .onAppear {
                     AppPresentation.settingsDidAppear()
                 }